Layouts of existing nodes are now re-used.
The ILayoutNodes in input parameter layoutNodes of method Layout are not
the same as the ILayoutNodes in oldLayout.layoutResult. We will get
always fresh ILayoutNodes in each call to this method. That is why
we need retrieve the old positions from oldLayout via the ID.

The code now works for the initial graph layout, but not really for
every other, because we assign positions to new nodes arbitrarily.
Leaf nodes and inner nodes must be scaled properly and much more.
Also the retrieval of nodes from the older layout is not efficient.
There is still a lot to do. I added a few FIXME markers, but there
is even more to it. This commit is only to show the root cause of
the problem we experienced and how it could be handled.

Cleaned up the code a bit.